remove tests artifacts
